src/requirements.txt
asgiref==3.8.1
Django==5.1
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
djangorestframework==3.15.2
django-passwords==0.3.12
django-sslserver==0.22
